+++ src/components/main/CategorySection/index.tsx
@@ -1,7 +1,7 @@
 import { useRouter } from 'next/router';
 import { useGetRoadmaps } from '@/api/hooks/roadmap';
 import Collapsible from '@/components/main/Collapsible';
-import { useParamsToUrl } from '@/hooks/common/useParamsToUrl';
+import { useParamsToUrl } from '@/hooks/useParamsToUrl';
 import * as Styled from './style';
 
 const CategorySection = () => {
