+++ src/components/main/History/index.tsx
@@ -5,7 +5,7 @@ import { ResponsiveCalendar } from '@nivo/calendar';
 import { useGetUserHistory } from '@/api/hooks/user';
 import CustomSuspense from '@/components/common/CustomSuspense';
 import Skeleton from '@/components/common/Skeleton';
-import { useParamsToUrl } from '@/hooks/common/useParamsToUrl';
+import { useParamsToUrl } from '@/hooks/useParamsToUrl';
 import * as Styled from './style';
 
 const History = () => {
